pyproject.toml
[project]
name = "pfruleslawyer"
version = "0.1.0"
description = "RAG-powered Pathfinder 1e rules assistant"
authors = [
    {name = "Jess Hyde",email = "jhyde369@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.15"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "html2text (>=2025.4.15,<2026.0.0)",
    "anthropic (>=0.40.0,<1.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "chromadb (>=0.5.0,<1.0.0)",
    "sentence-transformers (>=3.0.0,<4.0.0)",
    "spacy (>=3.7.0,<4.0.0)",
    "nltk (>=3.8.0,<4.0.0)",
    "posthog (>=2.4.0,<6.0.0)",
    "markdownify (>=1.2.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{include = "pfruleslawyer", from = "src"}]

[tool.poetry.scripts]
pfrules = "cli.ask:main"
pfrules-preprocess = "cli.preprocess:main"
pfrules-vectordb = "cli.vectordb:main"
pfrules-db = "cli.db:main"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"

[tool.pytest.ini_options]
pythonpath = ["src", "."]
testpaths = ["tests"]
